Move io methods to separate module
